+++ aquascope/collectors/__init__.py
@@ -3,6 +3,7 @@
 from aquascope.collectors.aquastat import AquastatCollector
 from aquascope.collectors.base import BaseCollector
 from aquascope.collectors.bom import BOMCollector
+from aquascope.collectors.brazil_ana import BrazilANACollector
 from aquascope.collectors.camels_br import CAMELSBRCollector
 from aquascope.collectors.camels_cl import CAMELSCLCollector
 from aquascope.collectors.copernicus import CopernicusCollector
@@ -39,6 +40,7 @@ __all__ = [
     "AquastatCollector",
     "BaseCollector",
     "BOMCollector",
+    "BrazilANACollector",
     "CAMELSCLCollector",
     "CAMELSBRCollector",
     "CopernicusCollector",